requirements.txt
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
audioread==3.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
ci_sdr==0.0.2
click==8.1.7
ConfigArgParse==1.7
ctc_segmentation==1.7.4
Cython==3.0.7
decorator==5.1.1
Distance==0.1.3
editdistance==0.6.2
einops==0.7.0
espnet==202310
espnet-tts-frontend==0.0.3
fast-bss-eval==0.1.3
filelock==3.13.1
fsspec==2023.12.2
g2p-en==2.1.0
h5py==3.10.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.6
importlib-metadata==4.13.0
inflect==7.0.0
jaconv==0.3.4
jamo==0.4.1
Jinja2==3.1.2
joblib==1.3.2
kaldiio==2.18.0
librosa==0.9.2
llvmlite==0.41.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numba==0.58.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==23.2
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.25.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyopenjtalk==0.3.3
pypinyin==0.44.0
pytorch-wpe==0.0.1
pyworld==0.3.4
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
resampy==0.4.2
scikit-learn==1.3.2
scipy==1.11.4
sentencepiece==0.1.97
six==1.16.0
soundfile==0.12.1
sympy==1.12
threadpoolctl==3.2.0
torch==2.1.2
torch-complex==0.4.3
torchaudio==2.1.2
tqdm==4.66.1
triton==2.1.0
typeguard==2.13.3
typing_extensions==4.9.0
Unidecode==1.3.7
urllib3==2.1.0
zipp==3.17.0
